Docker/Practice_3_Medium_Docs_Story_Uploader/requirements.txt
# The python modules which are required to be installed on the container before running the program itself.
python-docx
playwright